requirements.txt
click                  == 8.0.4
Flask                  == 2.0.3
Flask-Cors             == 3.0.10
itsdangerous           == 2.1.1
Jinja2                 == 3.0.3
MarkupSafe             == 2.1.0
mysql-connector-python == 8.0.28
pip                    == 21.2.3
protobuf               == 3.19.4
setuptools             == 57.4.0
six                    == 1.16.0
Werkzeug               == 2.0.3